Add GP-relative load and rework all other loads
GP-relative load was not getting picked up because LWs9_NM was catching
everything with LWs9_NM $r1, 0($r2). In order to avoid this kind of
behaviour, catching loads/stores that have an immediate which is out of
range had to be split into a separate pattern match (for all loads and
stores).

cherry-picks: 4df35217c304