NanoMIPS: Compress jump table entries
A custom lowering method for BR_JT ISDOpcode has been added. We're
emitting a MipsISD::BR_JT node with new pseudo instruction
LoadJumpTableOffset as a destination. For now, we're setting the entry
size to 1B for every jump table leaving the external assembler to decide
whether it can be done. We're also emitting relocation directive for
that purpose. Each jump table starts with a .jumptable directive.