Roll-back change to register pressure limit for nanoMIPS registers
The erstwhile default which assumes that the general purpose register
set is always under high pressure leads to better allocation on
average than the general MIPS model which allows free allocation of
up to 28 registers.  Revert to earlier behaviour for now, to be looked
in to along with scheduling changes.

Change-Id: I2e8b04edf49552a591f4f65cd33aef652a44e88e