Bug 87: Disable PC-relative addressing by default on assembler command-line
This brings LLVM behaviour in-line with GNU compiler in the following respect:
* PC-relative addressing is enabled by `.module pcrel' directive in assembly file
* The compiler will always generate the directive unless invoked with -mno-pcrel
* However when compiling hand-written assembly where the directive is not
present, the driver still passes -target-feature +pcrel to integrated assembler.
and LA macro is expanded to LAPC[48], instead of LI[48] as in the GNU compiler.

By not passing this target features to integrated assembler, we allow the
directive present (or absent) in the assembly file to control how the
object is generated. We still enable this target features by default
for the CodeGen step, so generated assembly will always contain the
requisite directives.

Change-Id: I6a4765d2d5db5201eb9a89cd297652dbd4f1c0bb