Bug102 follow-up: Skip label difference relocations for debug symbols
If symbols in debug sections (.debug_line/.debug_info) are not
anonymous, their differences are emitted as relocations. This causes
relocations to be emitted in the header region of debug sections. The
linker tries to interpret these sections before the relocations are
resolved and crashes.

Skip generation of relocations for symbols starting with
.Ldebug and .Lprologue which are emitted for debug sections.

Change-Id: I548f3fc5b95319b7163f9593322712a4b9316e03