[Bug90 1/2] Partial revert of commit d1994e5e0928367558090e513382a8e52c14fafa
This reverts the initial change that enabled round-section-sizes for
nanoMIPS by default.  Enabling this causes both the padding NOPs and
an addition _ALIGN relocation to be generated, which in hindsight
serves no purpose.

To contrast, GNU assembler generates the padding but not the
relocation by default. That too serves no real purpose and is infact
immune to clean-up by linker relaxation.

Change-Id: Idc29f64140568032d178840e8fb2da240246cf39