[Bug 90 2/2] Fix _ALIGN relocations to match GNU assembler
_ALIGN relocations for nanoMIPS must reference an absolute symbol
whose value resolves to (Log2 of) the requested alignment, unlike
RISCV which directly encodes a constant value in to the relocation
reference.  The size (ST_SIZE) field of the symbol should specify
the number of bytes of padding that assembler has already inserted
at the alignment point.

The earlier implementation mirrored RISCV which meant that the
linker could not re-adjust the alignment points during relaxation
leading to mis-aligned layout of code.

Change-Id: I11c3b3eebc14672cad36a2c27b3563308718c193